Refactor pyclaw/tests/test_io.py as to generate variants for serial and parallel hdf5 I/O tests
