+++ src/pyclaw/util.py
@@ -174,6 +174,16 @@ def gen_variants(application, verifier, kernel_languages=('Fortran',), disable_p
         yield test
     return
 
+def gen_io_variants(verifier, file_formats,regression_format, **kwargs):
+
+    arg_dicts = build_io_variant_arg_dicts(file_formats, regression_format)
+
+    for test_kwargs in arg_dicts:
+        test_kwargs.update(kwargs)
+        test = lambda: test_io(verifier, test_kwargs)
+        yield test
+    return
+
 def build_variant_arg_dicts(kernel_languages=('Fortran',), disable_petsc=False):
     import itertools
 
@@ -189,6 +199,15 @@ def build_variant_arg_dicts(kernel_languages=('Fortran',), disable_petsc=False):
     opt_names = 'use_petsc','kernel_language'
     opt_product = itertools.product(use_petsc_opts,kernel_languages)
     arg_dicts = [dict(zip(opt_names,argset)) for argset in opt_product]
+
+    return arg_dicts
+
+def build_io_variant_arg_dicts(file_formats=('hdf5','ascii'), regression_format='hdf5'):
+    opt_names = 'file_formats', 'regression_format'
+    opt_values = [(file_formats, regression_format)]
+    if regression_format=='hdf5':
+        opt_values.append((('hdf5p',), 'hdf5p'))
+    arg_dicts = [dict(zip(opt_names,argset)) for argset in opt_values]
 
     return arg_dicts
 
@@ -260,6 +279,20 @@ test error           : %s
         raise VerifyError(err)
     return
 
+def test_io(verifier, kwargs):
+    print kwargs
+
+    if 'regression_format' in kwargs and kwargs['regression_format'] != 'hdf5p':
+        try:
+            # don't duplicate serial test runs
+            from petsc4py import PETSc
+            rank = PETSc.COMM_WORLD.getRank()
+            if rank != 0:
+                return
+        except ImportError, e:
+            pass
+    return verifier(**kwargs)
+
 def check_diff(expected, test, **kwargs):
     r"""
     Checks the difference between expected and test values, return None if ok
@@ -284,8 +317,41 @@ def check_diff(expected, test, **kwargs):
     else:
         raise Exception('Incorrect use of check_diff verifier, specify tol!')
 
+def check_solutions_are_same(sol_a,sol_b):
+    assert len(sol_a.states) == len(sol_b.states)
+    assert sol_a.t == sol_b.t
+    for state in sol_a.states:
+        for ref_state in sol_b.states:
+            if ref_state.patch.patch_index == state.patch.patch_index:
+                break
 
+        # Required state attributes
+        assert np.linalg.norm(state.q - ref_state.q) < 1.e-6 # Not sure why this can be so large
+        if ref_state.aux is not None:
+            assert np.linalg.norm(state.aux - ref_state.aux) < 1.e-16
+        for attr in ['t', 'num_eqn', 'num_aux']:
+            assert getattr(state,attr) == getattr(ref_state,attr)
+        # Optional state attributes
+        for attr in ['patch_index', 'level']:
+            if hasattr(ref_state,attr):
+                assert getattr(state,attr) == getattr(ref_state,attr)
 
+        patch = state.patch
+        ref_patch = ref_state.patch
+        # Required patch attributes
+        for attr in ['patch_index', 'level']:
+            assert getattr(patch,attr) == getattr(ref_patch,attr)
+
+        dims = patch.dimensions
+        ref_dims = ref_patch.dimensions
+        for dim, ref_dim in zip(dims,ref_dims):
+            # Required dim attributes
+            for attr in ['num_cells','lower','delta']:
+                assert getattr(dim,attr) == getattr(ref_dim,attr)
+            # Optional dim attributes
+            for attr in ['units','on_lower_boundary','on_upper_boundary']:
+                if hasattr(ref_dim,attr):
+                    assert getattr(dim,attr) == getattr(ref_dim,attr)
 # ============================================================================
 #  F2PY Utility Functions
 # ============================================================================
